feat: 이미지 개수 검증 및 예외 처리
